third_party/WebKit/LayoutTests/svg/animations/animate-path-to-animation-expected.txt

SVG 1.1 dynamic animation tests
Test calcMode spline with to animation. You should see a green 100x100 path and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS path.getAttribute('d') is "M 40 40 L 60 40 L 60 60 L 40 60 Z"
PASS path.getAttribute('d') is "M 20 20 L 80 20 L 80 80 L 20 80 Z"
PASS path.getAttribute('d') is "M 0.00999928 0.00999928 L 99.99 0.00999928 L 99.99 99.99 L 0.00999928 99.99 Z"
PASS path.getAttribute('d') is "M 40 40 L 60 40 L 60 60 L 40 60 Z"
PASS successfullyParsed is true

TEST COMPLETE
